+++ app/Http/Controllers/OrdersController.php
@@ -65,6 +65,9 @@ class OrdersController extends Controller
         $order->user_id = $user->id;
         $order->save();
 
+        if(auth()->user()->id !== $user->orders->user_id)
+            return redirect('/home')->with('error', 'Você não possui permissão para fazer isso!');
+
         foreach($user->products as $product){
             $order->products()->attach($product->id, [
                 'product_quantity' => $product->pivot->product_quantity,
@@ -88,8 +91,25 @@ class OrdersController extends Controller
         $this->validate($request, [
             'status' => 'required|in:Cancelado,Em espera,Aprovado',
         ]);
-
+        
         $order = Order::find($id);
+
+        $permission = false;
+
+        if(auth()->user()->is_admin){
+            $permission = true;
+        } else {
+            foreach($order->users as $user){
+                if(auth()->user()->id === $user->id){
+                    $permission = true;
+                    break;
+                }
+            }
+        }        
+
+        if(!$permission)
+            return redirect('/home')->with('error', 'Você não possui permissão para fazer isso!');
+
         $order->status = $request->input('status');
         $order->save();
         
@@ -104,6 +124,9 @@ class OrdersController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(auth()->user()->id !== $user->orders->user_id && !auth()->user()->is_admin)
+            return redirect('/home')->with('error', 'Você não possui permissão para fazer isso!');
+
         $order = Order::find($request->id);
         $order->delete();
